Do not show drop message when user not connected
